files.html: Add links to table rows
Earlier we used <a> tags inside the table. This is not nice
as we have to click on the text to be able to go to the next
file. Now, we add a JS click event on the table row which
will automatically click the first <a> inside it for us.

Related to https://github.com/coala-analyzer/coala-html/issues/16