Use codecov-python
Currently codecov-bash is used, which involves an uncached fetch
of a unversioned bash script.
Using codecov-python allows the tool to be cached, and greater
reliability due to versioned releases of the tool.

In addition, codecov-bash does not work on Windows, and our
developer base is better tuned to investigate any python
related problems.

Fixes https://github.com/coala/coala-html/issues/117